linux

+++ src/main/java/dao/UsuarioDAO.java
@@ -18,17 +18,18 @@ public class UsuarioDAO {
 		connection = SingleConnectionBanco.getConnection();
 	}
 	
-	public ModelLogin gravarUsuario(ModelLogin objeto) throws Exception {
+	public ModelLogin gravarUsuario(ModelLogin objeto, Long userLogado) throws Exception {
 		
 		if(objeto.isNovo()) {
 		
-		String sql = "INSERT INTO model_login (nome, email, login, senha) VALUES (?, ?, ?, ?);";
+		String sql = "INSERT INTO model_login (nome, email, login, senha, usuario_id) VALUES (?, ?, ?, ?,?);";
 		
 		PreparedStatement preparedSQL = connection.prepareStatement(sql);
 		preparedSQL.setString(1, objeto.getNome());
 		preparedSQL.setString(2, objeto.getEmail());
 		preparedSQL.setString(3, objeto.getLogin());
 		preparedSQL.setString(4, objeto.getSenha());
+		preparedSQL.setLong(5, userLogado);
 		preparedSQL.execute();
 		
 		connection.commit();
@@ -45,14 +46,14 @@ public class UsuarioDAO {
 				connection.commit();
 				
 			}
-			return this.consultaUsuario(objeto.getLogin());
+			return this.consultaUsuario(objeto.getLogin(),userLogado);
 		
 	}
 	
-	public List<ModelLogin> consultaUsuarioList() throws Exception{
+	public List<ModelLogin> consultaUsuarioList(Long userLogado) throws Exception{
 		
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
-		String sql = "select * from model_login ;";
+		String sql = "select * from model_login where useradmin is false and usuario_id = "+userLogado+ ";";
 		
 		PreparedStatement statement = connection.prepareStatement(sql);
 		ResultSet resultado = statement.executeQuery();
@@ -71,11 +72,12 @@ public class UsuarioDAO {
 		return retorno;
 	}
 	
-	public List<ModelLogin> consultaUsuarioList(String nome) throws Exception{
+	public List<ModelLogin> consultaUsuarioList(String nome, Long userLogado) throws Exception{
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
-		String sql = "select * from model_login where nome like ? ;";
+		String sql = "select * from model_login where nome like ?  and useradmin is false  and usuario_id = ?;";
 		PreparedStatement statement = connection.prepareStatement(sql);
 		statement.setString(1, "%" +nome+ "%");
+		statement.setLong(2, userLogado);
 		ResultSet resultado = statement.executeQuery();
 		
 		while(resultado.next()) {
@@ -91,14 +93,55 @@ public class UsuarioDAO {
 		
 		return retorno;
 	}
-	
-	
-	
-	public ModelLogin consultaUsuario (String login) throws Exception {
+public ModelLogin consultaUsuarioLogado (String login) throws Exception {
 		
 		ModelLogin modelLogin = new ModelLogin();
 		
-		String sql = "select * from model_login where login = '"+login+"'";
+		String sql = "select * from model_login where login = '"+login+"';";
+		
+		PreparedStatement statement = connection.prepareStatement(sql);
+		
+		ResultSet resultado = statement.executeQuery();
+		
+		while (resultado.next()) {
+			modelLogin.setId(resultado.getLong("id"));
+			modelLogin.setEmail(resultado.getString("email"));
+			modelLogin.setLogin(resultado.getString("login"));
+			modelLogin.setNome(resultado.getString("nome"));
+			modelLogin.setSenha(resultado.getString("senha"));
+			modelLogin.setUseradmin(resultado.getBoolean("useradmin"));
+		}
+		
+		return modelLogin;
+	}
+	
+public ModelLogin consultaUsuario (String login) throws Exception {
+		
+		ModelLogin modelLogin = new ModelLogin();
+		
+		String sql = "select * from model_login where login = '"+login+"' and useradmin is false ;";
+		
+		PreparedStatement statement = connection.prepareStatement(sql);
+		
+		ResultSet resultado = statement.executeQuery();
+		
+		while (resultado.next()) {
+			modelLogin.setId(resultado.getLong("id"));
+			modelLogin.setEmail(resultado.getString("email"));
+			modelLogin.setLogin(resultado.getString("login"));
+			modelLogin.setNome(resultado.getString("nome"));
+			modelLogin.setSenha(resultado.getString("senha"));
+			modelLogin.setUseradmin(resultado.getBoolean("useradmin"));
+		}
+		
+		return modelLogin;
+	}
+	
+	public ModelLogin consultaUsuario (String login, Long userLogado) throws Exception {
+		
+		ModelLogin modelLogin = new ModelLogin();
+		
+		String sql = "select * from model_login where login = '"+login+"' and useradmin is false and usuario_id ="+userLogado+";";
 		
 		PreparedStatement statement = connection.prepareStatement(sql);
 		
@@ -114,14 +157,15 @@ public class UsuarioDAO {
 		
 		return modelLogin;
 	}
-	public ModelLogin consultaUsuarioId (String id) throws Exception {
+	public ModelLogin consultaUsuarioId (String id, Long userLogado) throws Exception {
 			
 			ModelLogin modelLogin = new ModelLogin();
 			
-			String sql = "select * from model_login where id = ? ;";
+			String sql = "select * from model_login where id = ? and useradmin is false;";
 			
 			PreparedStatement statement = connection.prepareStatement(sql);
 			statement.setLong(1, Long.parseLong(id));
+			statement.setLong(2, userLogado);
 			
 			ResultSet resultado = statement.executeQuery();
 			
@@ -151,7 +195,7 @@ public class UsuarioDAO {
 	
 	
 	public void deletarUsuario(String idUsuario) throws Exception {
-		String sql = "delete from model_login where id = ? ;";
+		String sql = "delete from model_login where id = ? and useradmin is false;";
 		PreparedStatement prepareSql = connection.prepareStatement(sql);
 		prepareSql.setLong(1, Long.parseLong(idUsuario));
 		prepareSql.executeUpdate();
